Add files via upload
Me and my supporters have decided the gift that will be given to our honorable heroes duelists, and the gift is whenever the those archtypes (will be mentioned below) will take a new booster pack, they will be added by 2 cards and they will decide how those cards will look. this is about the gift. 

and we have new 2 scums they decided to be punished by their own will, and I will view their IDs

--47198668--  AKA REIJI AKABA
He violated the rules which is "DON'T VIOLATE THE MODS" "DON'T ARGUE WITH A MOD ABOUT WHEN YOUR BAN WILL GET CANCELED" "NO SEXUAL AND VIOLENCE AND NO TERRORISM FOR THE SAKE OF THE TRADING CARD GAME"
FIRSTLY, HE WILL BE BANNED AS A DUELIST AND THE NEXT WORLDWIDE TOURNMET HE WON'T PARTICIPATE, IF SO, HIS ARCHTYPE WILL BE DELETED

--70902743-- AKA JACK ATLAS
THE SAME AS HIS FRIEND BUT HE DIDNT DO THE LAST RULE AND HE WILL GET THE SAME PUNISHMENT + HE LIED


THEY WILL BE ADDED TO THE SERVER WHICH CALLED "UNTITLED"

GOODLUCK YOU 2 SCUMS

SPECIAL THANKS TO OUR 2 HONORABLE HEROES DUELISTS